+++ src/orchard/inspect/analytics.clj
@@ -169,6 +169,20 @@
                        (let [kcoll (mapv #(get % k) coll)]
                          [k (basic-list-stats kcoll false)])))))))
 
+(defn- string-stats [^String s]
+  (when (string? s)
+    (let [len (.length s)]
+      (loop [i 0, non-ascii 0, non-bmp 0, lines 1]
+        (if (< i len)
+          (let [cp (.codePointAt s i)]
+            (recur (+ i (Character/charCount cp))
+                   (cond-> non-ascii (> cp 127) inc)
+                   (cond-> non-bmp (> cp 0xFFFF) inc)
+                   (cond-> lines (= cp 10) inc)))
+          {:non-ascii non-ascii
+           :non-bmp   non-bmp
+           :lines     lines})))))
+
 (defn analytics
   "Return various analytical data about `object`. Supports the following data
   types with different amount of insights:
@@ -186,6 +200,7 @@
     (or (tuples-stats object)
         (records-stats object)
         (keyvals-stats object)
+        (string-stats object)
         (basic-list-stats object true))))
 
 (defn can-analyze?
@@ -194,4 +209,5 @@
   (or (instance? List object)
       (instance? Map object)
       (instance? Set object)
+      (instance? String object)
       (some-> (class object) (.isArray))))